Minor reorg: Moved flickr functions out of main.js and into flickr.js
